fix frame resize issue / update anvil
