+++ apps/sledge/src-tauri/capabilities/default.json
@@ -2,7 +2,13 @@
   "$schema": "../gen/schemas/desktop-schema.json",
   "identifier": "default",
   "description": "enables the default permissions",
-  "windows": ["start", "editor-*", "about", "settings", "error"],
+  "windows": [
+    "start",
+    "editor-*",
+    "about",
+    "settings",
+    "error"
+  ],
   "permissions": [
     "core:default",
     "dialog:default",
@@ -53,6 +59,7 @@
     "core:webview:allow-set-webview-zoom",
     "os:default",
     "process:default",
-    "updater:default"
+    "updater:default",
+    "log:default"
   ]
 }
