+++ package.json
@@ -1,6 +1,6 @@
 {
   "private": true,
-  "packageManager": "pnpm@10.16.1",
+  "packageManager": "pnpm@10.18.1",
   "scripts": {
     "dev": "pnpm -F sledge dev",
     "build": "pnpm -F sledge build",
